Feature/ws improvements

+++ Client/src/app/content/folders/state/folders-state.ts
@@ -16,7 +16,7 @@ import { ApiFoldersService } from '../api-folders.service';
 import { FullFolder } from '../models/full-folder.model';
 import {
   LoadFolders,
-  AddFolder,
+  CreateFolder,
   SelectIdFolder,
   UnSelectIdFolder,
   UnSelectAllFolder,
@@ -38,10 +38,12 @@ import {
   GetInvitedUsersToFolder,
   AddToDomFolders,
   ResetFolders,
+  AddFolders,
 } from './folders-actions';
 import { Folders } from '../models/folders.model';
 import { InvitedUsersToNoteOrFolder } from '../../notes/models/invited-users-to-note.model';
 import { UpdateFolderUI } from './update-folder-ui.model';
+import { Router } from '@angular/router';
 
 interface FullFolderState {
   isOwner: boolean;
@@ -74,7 +76,10 @@ interface FolderState {
 })
 @Injectable()
 export class FolderStore {
-  constructor(private api: ApiFoldersService, private orderService: OrderService) {}
+  constructor(
+    private api: ApiFoldersService,
+    private orderService: OrderService,
+    private router: Router) {}
 
   static getFoldersByTypeStatic(state: FolderState, type: FolderTypeENUM) {
     return state.folders.find((x) => x.typeFolders === type);
@@ -229,22 +234,22 @@ export class FolderStore {
   @Action(DeleteFoldersPermanently)
   async deleteFoldersPermanently(
     { getState, dispatch, patchState }: StateContext<FolderState>,
-    { selectedIds }: DeleteFoldersPermanently,
+    { selectedIds, isCallApi }: DeleteFoldersPermanently,
   ) {
-    await this.api.deleteFolders(selectedIds).toPromise();
 
-    const foldersFrom = this.getFoldersByType(getState, FolderTypeENUM.Deleted);
-    const foldersFromNew = foldersFrom.filter((x) => this.itemNoFromFilterArray(selectedIds, x));
-    dispatch(
-      new UpdateFolders(
-        new Folders(FolderTypeENUM.Deleted, foldersFromNew),
-        FolderTypeENUM.Deleted,
-      ),
-    );
+    if(isCallApi){
+      await this.api.deleteFolders(selectedIds).toPromise();
+    }
+
+    for (const { folders, typeFolders } of getState().folders) {
+      const foldersFromNew = folders.filter((x) => this.itemNoFromFilterArray(selectedIds, x));
+      dispatch(new UpdateFolders(new Folders(typeFolders, foldersFromNew), typeFolders));
+    }
 
     patchState({
       removeFromMurriEvent: [...selectedIds],
     });
+
     dispatch([UnSelectAllFolder, RemoveFromDomMurri]);
   }
 
@@ -444,12 +449,20 @@ export class FolderStore {
   }
 
   // FUNCTIONS
-  @Action(AddFolder)
+  @Action(CreateFolder)
   async newFolder({ getState, dispatch }: StateContext<FolderState>) {
     const newF = await this.api.new().toPromise();
     const folders = this.getFoldersByType(getState, FolderTypeENUM.Private);
     const toUpdate = new Folders(FolderTypeENUM.Private, [newF, ...folders]);
     dispatch(new UpdateFolders(toUpdate, FolderTypeENUM.Private));
+    this.router.navigate([`folders/${newF.id}`])
+  }
+
+  @Action(AddFolders)
+  addFolder({ getState, dispatch }: StateContext<FolderState>, { folders, type } : AddFolders) {
+    const foldersState = this.getFoldersByType(getState, type);
+    const toUpdate = new Folders(type, [...folders, ...foldersState]);
+    dispatch(new UpdateFolders(toUpdate, type));
   }
 
   @Action(ChangeColorFolder)
