+++ Client/src/app/content/notes/state/notes-state.ts
@@ -15,7 +15,7 @@ import { ShowSnackNotification } from 'src/app/core/stateApp/app-action';
 import { ApiServiceNotes } from '../api-notes.service';
 import {
   LoadNotes,
-  AddNote,
+  CreateNote,
   ChangeColorNote,
   SelectIdNote,
   UnSelectIdNote,
@@ -46,6 +46,7 @@ import {
   LoadSnapshotNote,
   ResetNotes,
   ChangeTypeNote,
+  AddNotes,
 } from './notes-actions';
 import { UpdateNoteUI } from './update-note-ui.model';
 import { SmallNote } from '../models/small-note.model';
@@ -60,6 +61,8 @@ import { ApiNoteHistoryService } from '../full-note/services/api-note-history.se
 import { ApiTextService } from '../full-note/services/api-text.service';
 import { LongTermOperationsHandlerService } from '../../long-term-operations-handler/services/long-term-operations-handler.service';
 import { LongTermsIcons } from '../../long-term-operations-handler/models/long-terms.icons';
+import { Router } from '@angular/router';
+import { NoteSnapshot } from '../full-note/models/history/note-snapshot.model';
 
 interface FullNoteState {
   note: FullNote;
@@ -107,7 +110,8 @@ export class NoteStore {
     private orderService: OrderService,
     private historyApi: ApiNoteHistoryService,
     private longTermOperationsHandler: LongTermOperationsHandlerService,
-  ) {}
+    private router: Router
+  ) { }
 
   static getNotesByTypeStatic(state: NoteState, type: NoteTypeENUM) {
     return state.notes.find((x) => x.typeNotes === type);
@@ -215,12 +219,17 @@ export class NoteStore {
   static oneFull(state: NoteState): FullNote {
     return state.fullNoteState?.note;
   }
+  
+  @Selector()
+  static fullNoteTitle(state: NoteState): string {
+    return state.fullNoteState?.note?.title;
+  }
 
   @Selector()
   static canEdit(state: NoteState): boolean {
     return state.fullNoteState?.canEdit;
   }
-  
+
   @Selector()
   static canView(state: NoteState): boolean {
     return state.fullNoteState?.canView;
@@ -244,6 +253,16 @@ export class NoteStore {
   @Selector()
   static snapshotState(state: NoteState): NoteSnapshotState {
     return state.snapshotState;
+  }
+
+  @Selector()
+  static snapshotNote(state: NoteState): NoteSnapshot {
+    return state.snapshotState?.noteSnapshot;
+  }
+
+  @Selector()
+  static snapshotNoteTitle(state: NoteState): string {
+    return state.snapshotState?.noteSnapshot?.title;
   }
 
   // Get notes
@@ -316,14 +335,20 @@ export class NoteStore {
     return state.isCanceled;
   }
 
-  @Action(AddNote)
+  @Action(CreateNote)
   async newNote({ getState, dispatch }: StateContext<NoteState>) {
     const note = await this.api.new().toPromise();
-
     const notes = this.getNotesByType(getState, NoteTypeENUM.Private);
-
     const toUpdate = new Notes(NoteTypeENUM.Private, [note, ...notes]);
     dispatch(new UpdateNotes(toUpdate, NoteTypeENUM.Private));
+    this.router.navigate([`notes/${note.id}`]);
+  }
+
+  @Action(AddNotes)
+  addNote({ getState, dispatch }: StateContext<NoteState>, { notes, type } : AddNotes) {
+    const notesState = this.getNotesByType(getState, type);
+    const toUpdate = new Notes(type, [...notes, ...notesState]);
+    dispatch(new UpdateNotes(toUpdate, type));
   }
 
   @Action(UpdateNotes)
@@ -346,17 +371,23 @@ export class NoteStore {
   @Action(DeleteNotesPermanently)
   async deleteNotesPermanently(
     { getState, dispatch, patchState }: StateContext<NoteState>,
-    { selectedIds }: DeleteNotesPermanently,
+    { selectedIds, isCallApi }: DeleteNotesPermanently,
   ) {
-    await this.api.deleteNotes(selectedIds).toPromise();
 
-    const notesFrom = this.getNotesByType(getState, NoteTypeENUM.Deleted);
-    const notesFromNew = notesFrom.filter((x) => this.itemNoFromFilterArray(selectedIds, x));
-    dispatch(new UpdateNotes(new Notes(NoteTypeENUM.Deleted, notesFromNew), NoteTypeENUM.Deleted));
+    if (isCallApi) {
+      const resp = await this.api.deleteNotes(selectedIds).toPromise();
+      if(!resp.success){ return; }
+    }
+
+    for (const { notes, typeNotes } of getState().notes) {
+      const notesFromNew = notes.filter((x) => this.itemNoFromFilterArray(selectedIds, x));
+      dispatch(new UpdateNotes(new Notes(typeNotes, notesFromNew), typeNotes));
+    }
 
     patchState({
       removeFromMurriEvent: [...selectedIds],
     });
+
     dispatch([UnSelectAllNote, RemoveFromDomMurri]);
   }
 
@@ -479,10 +510,10 @@ export class NoteStore {
     if (resp.success) {
       const fullNote = getState().fullNoteState?.note;
       if (fullNote && selectedIds.some(id => id === fullNote.id)) {
-        patchState({ fullNoteState: { ...getState().fullNoteState, note: {...fullNote, color }}});
+        patchState({ fullNoteState: { ...getState().fullNoteState, note: { ...fullNote, color } } });
       }
       const notesForUpdate = this.getNotesByIds(getState, selectedIds);
-      if(notesForUpdate && notesForUpdate.length > 0) {
+      if (notesForUpdate && notesForUpdate.length > 0) {
         notesForUpdate.forEach((note) => note.color = color);
         const updatesUI = notesForUpdate.map((note) =>
           this.toUpdateNoteUI(note.id, note.color, null, null, null, null),
@@ -533,12 +564,12 @@ export class NoteStore {
     { label, selectedIds, isCallApi, errorPermissionMessage }: AddLabelOnNote,
   ) {
     let resp: OperationResult<any> = { success: true, data: null, message: null };
-    if (isCallApi){
+    if (isCallApi) {
       resp = await this.api.addLabel(label.id, selectedIds).toPromise();
     }
-    if(resp.success){
+    if (resp.success) {
       const note = getState().fullNoteState?.note;
-      if(note && selectedIds.some(id => id === note.id)){
+      if (note && selectedIds.some(id => id === note.id)) {
         patchState({ fullNoteState: { ...getState().fullNoteState, note: { ...note, labels: [...note.labels, label] } } });
       }
       // Updates small notes
@@ -575,16 +606,16 @@ export class NoteStore {
     { labelId, selectedIds, isCallApi, errorPermissionMessage }: RemoveLabelFromNote,
   ) {
     let resp: OperationResult<any> = { success: true, data: null, message: null };
-    if (isCallApi){
+    if (isCallApi) {
       resp = await this.api.removeLabel(labelId, selectedIds).toPromise();
     }
-    if(resp.success){
+    if (resp.success) {
       let note = getState().fullNoteState?.note;
-      if(note && selectedIds.some(id => id === note.id)){
+      if (note && selectedIds.some(id => id === note.id)) {
         note = { ...note, labels: note.labels.filter((z) => z.id !== labelId) };
-        patchState({ fullNoteState: { ...getState().fullNoteState, note }});
+        patchState({ fullNoteState: { ...getState().fullNoteState, note } });
       }
-      const notesForUpdate = this.getNotesByIds(getState, selectedIds);  
+      const notesForUpdate = this.getNotesByIds(getState, selectedIds);
       const updateNoteEvent: UpdateNoteUI[] = [];
       notesForUpdate.forEach((x: SmallNote) => {
         x.labels = x.labels.filter((z) => z.id !== labelId);
@@ -676,7 +707,7 @@ export class NoteStore {
 
   @Action(UpdateOneNote)
   updateOneSmallNote({ dispatch, getState }: StateContext<NoteState>, { note }: UpdateOneNote) {
-    for(const noteState of getState().notes) {
+    for (const noteState of getState().notes) {
       let isUpdate = false;
       const notes = noteState.notes.map((x) => {
         let nt = { ...x };
@@ -686,7 +717,7 @@ export class NoteStore {
         }
         return nt;
       });
-      if(isUpdate){
+      if (isUpdate) {
         dispatch(new UpdateNotes(new Notes(note.noteTypeId, [...notes]), note.noteTypeId));
       }
     }
@@ -739,10 +770,10 @@ export class NoteStore {
     if (resp.success) {
       const fullNote = getState().fullNoteState?.note;
       if (fullNote && fullNote.id === noteId) {
-        patchState({ fullNoteState: { ...getState().fullNoteState, note: {...fullNote, title: str} } });
+        patchState({ fullNoteState: { ...getState().fullNoteState, note: { ...fullNote, title: str } } });
       }
       const noteUpdate = this.getNoteById(getState, noteId);
-      if(noteUpdate){
+      if (noteUpdate) {
         noteUpdate.title = str;
         patchState({ updateNoteEvent: [this.toUpdateNoteUI(noteUpdate.id, null, null, null, null, noteUpdate.title)] });
         dispatch(new UpdateOneNote(noteUpdate));
@@ -879,8 +910,8 @@ export class NoteStore {
   }
 
   getNoteById = (getState: () => NoteState, id: string): SmallNote => {
-    for(const notes of getState().notes) {
-      for(const x of notes.notes){
+    for (const notes of getState().notes) {
+      for (const x of notes.notes) {
         const note = { ...x };
         if (id === note.id) {
           return note;
